🏗 Add dev-mode for `amp visual-diff` and split code into modules (#36673)
