+++ build-system/tasks/visual-diff/package.json
@@ -7,7 +7,11 @@
     "@percy/core": "1.0.0-beta.70",
     "@percy/puppeteer": "2.0.0",
     "@percy/sdk-utils": "1.0.0-beta.70",
+    "@types/inquirer": "8.1.3",
     "@types/puppeteer": "5.4.4",
-    "puppeteer": "10.4.0"
+    "inquirer": "8.2.0",
+    "puppeteer": "10.4.0",
+    "puppeteer-extra": "3.2.3",
+    "puppeteer-extra-plugin-user-preferences": "2.3.1"
   }
 }
