+++ examples/visual-tests/amp-date-picker/amp-date-picker.js
@@ -2,7 +2,7 @@
 
 const {
   verifySelectorsVisible,
-} = require('../../../build-system/tasks/visual-diff/helpers');
+} = require('../../../build-system/tasks/visual-diff/verifiers');
 
 module.exports = {
   'select a different date': async (page, name) => {
