+++ examples/visual-tests/amp-list/amp-list.amp.js
@@ -2,7 +2,7 @@
 
 const {
   verifySelectorsInvisible,
-} = require('../../../build-system/tasks/visual-diff/helpers');
+} = require('../../../build-system/tasks/visual-diff/verifiers');
 
 module.exports = {
   'tap "see more" button': async (page, name) => {
