+++ examples/visual-tests/amp-sidebar/amp-sidebar-toolbar.js
@@ -6,7 +6,7 @@
 
 const {
   verifySelectorsVisible,
-} = require('../../../build-system/tasks/visual-diff/helpers');
+} = require('../../../build-system/tasks/visual-diff/verifiers');
 
 module.exports = {
   'open sidebar toolbar': async (page, name) => {
