+++ examples/visual-tests/amp-story/amp-story-ad.js
@@ -2,7 +2,7 @@
 
 const {
   verifySelectorsVisible,
-} = require('../../../build-system/tasks/visual-diff/helpers');
+} = require('../../../build-system/tasks/visual-diff/verifiers');
 
 module.exports = {
   'Test story ad system layer property correctness': async (page, name) => {
