+++ examples/visual-tests/amp-story/amp-story-pagination-buttons.js
@@ -2,7 +2,7 @@
 
 const {
   verifySelectorsVisible,
-} = require('../../../build-system/tasks/visual-diff/helpers');
+} = require('../../../build-system/tasks/visual-diff/verifiers');
 
 module.exports = {
   'shows corresponding buttons for first page': async (page, name) => {
